backend_py/requirements.txt
# Python Backend for Floria App

Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
mssql-django>=1.4
python-dotenv>=1.0
bcrypt>=4.2
drf-spectacular>=0.27

# FastAPI Service Layer
fastapi>=0.115
uvicorn>=0.30
pydantic>=2.9
pyodbc>=5.1
PyJWT>=2.9
google-auth>=2.35
google-auth-oauthlib>=1.2
requests>=2.31

# Development
pytest>=8.0
pytest-django>=4.8

